Make suggested optional changes
